I added 	new file:   IMG30.JPG 	new file:   cam.py         new file:   fly.jpg         new file:   mario.jpg         new file:   pyramids.py         new file:   pyramids_2.py         new file:   res.png         new file:   temp.py         new file:   temp2.py         new file:   template.jpg for that I'm afraid that my computer get some risk, and I need to reinstall the OS.
These files are my first test to achieve the courese-to-fine template matching with python openCV.

Up to now, I could match the template and make pyramids for images.
But there are several points that they are too closed to each other.
I need to define a compare function to select the best ones.

 Initial commit